Refactors operator implementations to standalone classes [#819]
