+++ lang/src/org/partiql/lang/eval/physical/operators/AggregateOperatorFactory.kt
@@ -14,10 +14,27 @@
 package org.partiql.lang.eval.physical.operators
 
 import org.partiql.lang.domains.PartiqlPhysical
+import org.partiql.lang.eval.DEFAULT_COMPARATOR
+import org.partiql.lang.eval.ExprValue
+import org.partiql.lang.eval.physical.EvaluatorState
 import org.partiql.lang.eval.physical.SetVariableFunc
+import org.partiql.lang.eval.relation.RelationIterator
+import org.partiql.lang.eval.relation.RelationType
+import org.partiql.lang.eval.relation.relation
+import org.partiql.lang.planner.transforms.DEFAULT_IMPL_NAME
+import java.util.TreeMap
 
+/**
+ * Provides an implementation of the [PartiqlPhysical.Bexpr.Aggregate] operator.
+ *
+ * @constructor
+ *
+ * @param name
+ */
 public abstract class AggregateOperatorFactory(name: String) : RelationalOperatorFactory {
-    public final override val key: RelationalOperatorFactoryKey = RelationalOperatorFactoryKey(RelationalOperatorKind.AGGREGATE, name)
+
+    public override val key = RelationalOperatorFactoryKey(RelationalOperatorKind.AGGREGATE, name)
+
     public abstract fun create(
         source: RelationExpression,
         strategy: PartiqlPhysical.GroupingStrategy,
@@ -38,3 +55,58 @@ public class CompiledAggregateFunction(
     val value: ValueExpression,
     val quantifier: PartiqlPhysical.SetQuantifier,
 )
+
+internal object AggregateOperatorFactoryDefault : AggregateOperatorFactory(DEFAULT_IMPL_NAME) {
+    override fun create(
+        source: RelationExpression,
+        strategy: PartiqlPhysical.GroupingStrategy,
+        keys: List<CompiledGroupKey>,
+        functions: List<CompiledAggregateFunction>
+    ): RelationExpression = AggregateOperatorDefault(source, keys, functions)
+}
+
+internal class AggregateOperatorDefault(
+    val source: RelationExpression,
+    val keys: List<CompiledGroupKey>,
+    val functions: List<CompiledAggregateFunction>
+) : RelationExpression {
+    override fun evaluate(state: EvaluatorState): RelationIterator = relation(RelationType.BAG) {
+        val aggregationMap = TreeMap<ExprValue, List<Accumulator>>(DEFAULT_COMPARATOR)
+
+        val sourceIter = source.evaluate(state)
+        while (sourceIter.nextRow()) {
+
+            // Initialize the AggregationMap
+            val evaluatedGroupByKeys =
+                keys.map { it.value.invoke(state) }.let { state.valueFactory.newList(it) }
+            val accumulators = aggregationMap.getOrPut(evaluatedGroupByKeys) {
+                functions.map { function ->
+                    Accumulator.create(function.name, function.quantifier, state.valueFactory)
+                }
+            }
+
+            // Aggregate Values in Aggregation State
+            functions.forEachIndexed { index, function ->
+                val valueToAggregate = function.value(state)
+                accumulators[index].next(valueToAggregate)
+            }
+        }
+
+        // No Aggregations Created
+        if (keys.isEmpty() && aggregationMap.isEmpty()) {
+            functions.forEach { function ->
+                val accumulator = Accumulator.create(function.name, function.quantifier, state.valueFactory)
+                function.setAggregateVal(state, accumulator.compute())
+            }
+            yield()
+            return@relation
+        }
+
+        // Place Aggregated Values into Result State
+        aggregationMap.forEach { (exprList, accumulators) ->
+            exprList.forEachIndexed { index, exprValue -> keys[index].setGroupKeyVal(state, exprValue) }
+            accumulators.forEachIndexed { index, acc -> functions[index].setAggregateVal(state, acc.compute()) }
+            yield()
+        }
+    }
+}
