+++ lang/src/org/partiql/lang/eval/physical/operators/LetRelationalOperatorFactory.kt
@@ -1,30 +1,64 @@
 package org.partiql.lang.eval.physical.operators
 
 import org.partiql.lang.domains.PartiqlPhysical
-import org.partiql.lang.eval.physical.SetVariableFunc
+import org.partiql.lang.eval.physical.EvaluatorState
+import org.partiql.lang.eval.physical.VariableBinding
+import org.partiql.lang.eval.relation.RelationIterator
+import org.partiql.lang.eval.relation.relation
+import org.partiql.lang.planner.transforms.DEFAULT_IMPL_NAME
 
-/** A compiled variable binding. */
-class VariableBinding(
-    /** The function to be invoked at evaluation-time to set the value of the variable. */
-    val setFunc: SetVariableFunc,
-    /** The function to be invoked at evaluation-time to compute the value of the variable. */
-    val expr: ValueExpression
-)
-
-/** Provides an implementation of the [PartiqlPhysical.Bexpr.Let] operator.*/
+/**
+ * Provides an implementation of the [PartiqlPhysical.Bexpr.Let] operator.
+ *
+ * @constructor
+ *
+ * @param name
+ */
 abstract class LetRelationalOperatorFactory(name: String) : RelationalOperatorFactory {
-    final override val key: RelationalOperatorFactoryKey = RelationalOperatorFactoryKey(RelationalOperatorKind.LET, name)
 
-    /** Creates a [RelationExpression] instance for [PartiqlPhysical.Bexpr.Let]. */
+    final override val key = RelationalOperatorFactoryKey(RelationalOperatorKind.LET, name)
+
+    /**
+     * Creates a [RelationExpression] instance for [PartiqlPhysical.Bexpr.Let].
+     *
+     * @param impl
+     * @param sourceBexpr
+     * @param bindings list of [VariableBinding]s in the `LET` clause
+     * @return
+     */
     abstract fun create(
-        /**
-         * Contains any static arguments needed by the operator implementation that were supplied by the planner
-         * pass which specified the operator implementation.
-         */
         impl: PartiqlPhysical.Impl,
-        /** Invokes the source bindings expression. */
         sourceBexpr: RelationExpression,
-        /** List of [VariableBinding]s that were included in the `LET` clause. */
         bindings: List<VariableBinding>
     ): RelationExpression
+}
+
+internal object LetRelationalOperatorFactoryDefault : LetRelationalOperatorFactory(DEFAULT_IMPL_NAME) {
+
+    override fun create(
+        impl: PartiqlPhysical.Impl,
+        sourceBexpr: RelationExpression,
+        bindings: List<VariableBinding>
+    ) = LetOperator(
+        input = sourceBexpr,
+        bindings = bindings,
+    )
+}
+
+internal class LetOperator(
+    private val input: RelationExpression,
+    private val bindings: List<VariableBinding>
+) : RelationExpression {
+
+    override fun evaluate(state: EvaluatorState): RelationIterator {
+        val rows = input.evaluate(state)
+        return relation(rows.relType) {
+            while (rows.nextRow()) {
+                bindings.forEach {
+                    it.setFunc(state, it.expr(state))
+                }
+                yield()
+            }
+        }
+    }
 }
